remove lock structure from gpu feature
